Kotlin DSA Practice - restructed and renamed Added a new problem under string package.
