+++ src/main/kotlin/blind75_problems_by_category/Main.kt
@@ -1,3 +1,5 @@
+package blind75_problems_by_category
+
 fun main(args: Array<String>) {
     println("Hello World!")
