+++ src/main/kotlin/blind75_problems_by_category/array/LeetCode_ContainsDuplicate.kt
@@ -1,4 +1,4 @@
-package based_on_data_structure.array
+package blind75_problems_by_category.array
 
 /**
  * 217. Contains Duplicate
