+++ src/main/kotlin/blind75_problems_by_category/array/LeetCode_MaximumSubarray.kt
@@ -1,4 +1,4 @@
-package based_on_data_structure.array
+package blind75_problems_by_category.array
 
 /**
  * 53. Maximum Subarray
