+++ src/main/kotlin/blind75_problems_by_category/string/LeetCode_ValidAnagram.kt
@@ -1,4 +1,4 @@
-package based_on_data_structure.string
+package blind75_problems_by_category.string
 
 /**
  * 242. Valid Anagram
